wget: fix fetching of https URLs with http proxy
When fetching a https:// URL with HTTP proxy configured
(e.g. with environment variable http_proxy=http://your-proxy:3128)
busybox was making a https connection to the proxy.

This was because the protocol scheme for the target URL was used to determine whether to connect to the proxy over SSL or not.

When the proxy is in use, the decision on whether to connect to the proxy over https
should based on the proxy URL not on the target URL.

function                                             old     new   delta
wget_main                                           2381    2387      +6

Signed-off-by: Peter Lloyd <l-busybox@pgl22.co.uk>
Signed-off-by: Denys Vlasenko <vda.linux@googlemail.com>